move motor and accel to 0.5[m/s] can move

+++ Core/Inc/inc/PID.h
@@ -18,7 +18,7 @@ class PosType_PID : virtual private I_PID{
     
 public:
     PosType_PID(float _period = 1.0f, float _p_gain = 0, float _i_gain = 0, float _d_gain = 0):I_PID( _period, _p_gain, _i_gain, _d_gain) {}
-    float operator()(int);
+    float operator()(float);
 };
 
 
@@ -26,5 +26,5 @@ class SpeedType_PID : virtual private I_PID{
     
 public:
     SpeedType_PID(float _period = 1.0f, float _p_gain = 0, float _i_gain = 0, float _d_gain = 0):I_PID(_period ,_p_gain, _i_gain, _d_gain){}
-    float operator()(int);
+    float operator()(float);
 };
